@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "bootstrap": "wpackio-scripts bootstrap",
-    "build": "wpackio-scripts build",
+    "build": "export NODE_OPTIONS=--openssl-legacy-provider && wpackio-scripts build",
     "start": "wpackio-scripts start",
     "archive": "wpackio-scripts pack"
   },
@@ -22,12 +22,14 @@
     "@wpackio/entrypoint": "^6.4.0",
     "axios": "^0.21.4",
     "cupertino-pane": "^1.3.0",
-    "leaflet": "^1.7.1",
+    "leaflet": "^1.8.x",
     "leaflet-gesture-handling": "^1.2.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "npm-check-updates": "^16.14.20",
+    "react": "^18.0.x",
+    "react-dom": "^18.0.x",
     "react-hot-loader": "^4.13.0",
-    "react-leaflet": "^3.2.5",
+    "react-leaflet": "^4.2.x",
+    "react-leaflet-cluster": "^2.1.0",
     "sass": "^1.42.1",
     "turf": "^3.0.14",
     "url-loader": "^4.1.1",
