@@ -3,7 +3,7 @@
  * Plugin Name: CP Locations
  * Plugin URL: https://churchplugins.com
  * Description: Church locations plugin for managing campuses
- * Version: 1.0.10.1
+ * Version: 1.0.10.2
  * Author: Church Plugins
  * Author URI: https://churchplugins.com
  * Text Domain: cp-locations
